Fix Debug Build Error in Webcore module. https://bugs.webkit.org/show_bug.cgi?id=140383
Patch by Shivakumar JM <shiva.jm@samsung.com> on 2015-01-13
Reviewed by Csaba Osztrogonác.

No new tests because there is no visible behavior change.

* platform/linux/MemoryPressureHandlerLinux.cpp:
(WebCore::MemoryPressureHandler::ReliefLogger::platformLog):

Canonical link: https://commits.webkit.org/158368@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178349 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/linux/MemoryPressureHandlerLinux.cpp
@@ -209,11 +209,11 @@ void MemoryPressureHandler::ReliefLogger::platformLog()
 
     ssize_t memoryDiff = currentMemory - m_initialMemory;
     if (memoryDiff < 0)
-        LOG(MemoryPressure, "Pressure relief: %s: -dirty %ld bytes (from %ld to %ld)", m_logString, (memoryDiff * -1), m_initialMemory, currentMemory);
+        LOG(MemoryPressure, "Pressure relief: %s: -dirty %lu bytes (from %lu to %lu)", m_logString, static_cast<unsigned long>(memoryDiff * -1), static_cast<unsigned long>(m_initialMemory), static_cast<unsigned long>(currentMemory));
     else if (memoryDiff > 0)
-        LOG(MemoryPressure, "Pressure relief: %s: +dirty %ld bytes (from %ld to %ld)", m_logString, memoryDiff, m_initialMemory, currentMemory);
+        LOG(MemoryPressure, "Pressure relief: %s: +dirty %lu bytes (from %lu to %lu)", m_logString, static_cast<unsigned long>(memoryDiff), static_cast<unsigned long>(m_initialMemory), static_cast<unsigned long>(currentMemory));
     else
-        LOG(MemoryPressure, "Pressure relief: %s: =dirty (at %ld bytes)", m_logString, currentMemory);
+        LOG(MemoryPressure, "Pressure relief: %s: =dirty (at %lu bytes)", m_logString, static_cast<unsigned long>(currentMemory));
 }
 
 size_t MemoryPressureHandler::ReliefLogger::platformMemoryUsage()
